Move auth_method
