+++ auth_backend/auth_plugins/vk.py
@@ -10,6 +10,9 @@ from fastapi.background import BackgroundTasks
 from fastapi_sqlalchemy import db
 from pydantic import BaseModel, Field
 
+from auth_backend.auth_method.auth_method import AuthMethodMeta
+from auth_backend.auth_method.oauth import OauthMeta
+from auth_backend.auth_method.session import Session
 from auth_backend.exceptions import AlreadyExists, OauthAuthFailed
 from auth_backend.kafka.kafka import get_kafka_producer
 from auth_backend.models.db import User, UserSession
@@ -18,7 +21,6 @@ from auth_backend.utils.security import UnionAuth
 from auth_backend.utils.string import concantenate_strings
 
 from ..schemas.types.scopes import Scope
-from .auth_method import AuthMethodMeta, OauthMeta, Session
 
 
 logger = logging.getLogger(__name__)
